Move helper functions to main package
